Tyrian supplies canvas and context
