+++ indigo-render/src/indigo/render/RendererConfig.scala
@@ -6,7 +6,5 @@ final class RendererConfig(
     val clearColor: RGBA,
     val magnification: Int,
     val maxBatchSize: Int,
-    val antiAliasing: Boolean,
-    val premultipliedAlpha: Boolean,
     val transparentBackground: Boolean
 )
